requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-huggingface
langchain-text-splitters
pinecone-client
transformers
sentence-transformers
pillow
faiss-cpu
python-dotenv
langgraph
tavily-python
langchain-google-genai
unstract
sentence-transformers
transformers
torch
tqdm
uuid
